Added pure functions for type partial ordering
The typechecker is dependent on subtyping, but the typechecker's subtype
function is linked into the typechecker monad and Gamma. The
PartialOrder module implements subtyping by making Type a member of the
PartialOrd typeclass. I also add helper functions for finding the
minima and maxima; the nearest subtype; and subtype boolean function.

The immediate use of this subtyping is to find the most specific
serialization functions.

TODO:
 [ ] write unit and/or property tests
 [ ] use `mostSpecificSubtypes` to find serialization functions
 [ ] write nexus and pool stub code generators
 [ ] iteratively improve the stubs until they aren't stubs
 [ ] write tests for currying
 [ ] push to master