(3 undefined) push code generation to last step
I finally have a good diagram for code generation (scribbled in a
notebook). Working through the diagram, I realized my implementation was
still misformed. Encoding the pools should come after segmenting them
into pools.

In this commit, I pull all high order logic into the generate function.
It steps through the entire algorithm:

 1. findSerializers
 2. find roots
 3. for each root build a multi-lingual bitree
 4. for each bitree, collapse each node to a single language
 5. generate nexus code
 6. segment bitree into single language trees
 7. group segments into single language pools
 8. translate the pools into code

Besides changing the order (moving translation to the end) I also add
explicit Foreign nodes in step 6. When a foreign call is made, two nodes
are generated. One is a leaf is the first tree that is labeled Foreign.
The other is the root of a second tree in the foreign language.

The functions used in steps 6, 7, and 8 are stubs.