(7 fail) found problem, seek solution, break all
The core structure of my code generator, SAnno, is not general enough. I
originally thought that it would be sufficient to store the type and
expression for every node in the general tree. So I defined the pair of
types:

``` haskell
data SAnno a = SAnno (SExpr a) a

data SExpr a
    = VarS EVar
    | ListS [SAnno a]
    ...
    | AppS (SAnno a) [SAnno a]
    ...
    Foreigns Int Lang
```

This mutually recursive pair of types define a bipartite tree that
couples a generic annotation to each node. Prior to realization, this
annotation stores `[Type]`, afterwards it stores `Type`. That is, I
thought that by just mapping over the tree, I could translate to the
final concrete tree.

However, different languages may have different trees. This means I need
to further generalize SAnno, maybe to this:

``` haskell
data SAnno' f a b = SAnno (f (SExpr f a b), a) b

data SExpr f a b
    = VarS EVar
    | ListS [SAnno f a b]
    ...
    | AppS (SAnno a b) [SAnno f a b]
    ...
    Foreigns Int Lang

data Mono a = Mono a
```

 * `f` is a collection type (e.g., Set or Mono)
 * `a` is the annotation for the expression (e.g., the concrete type)
 * `b` is the general annotation (e.g., the general type)

The realization step transforms from `f::Set` to `f::Mono`.