(2 fail) - implement realization optimizer
My previous optimization "algorithm" simply chose the first tree. Now I
use a interop cost matrix and find the tree that minimizes interop
costs.

For example:

```
    R       Py     C
 R  100     50000  100
 Py 1000000 10     100
 C  1000000 50000  1
```

This is a mostly arbitrary matrix. Currently the cost of calling the R
interpreter, and loading required libraries, dwarfs all within language
calls and is far greater than the python cost as well. Thus optimization
is basically the minimization of calls to interpreted languages. As the
language interop methods are improved, this matrix will change.

I am also currently not considering the cost of running the functions
themselves. Eventually I can add performance models as either
benchmark-generated or given equations in the function metadata. Then I
can model the performance of the entire program which will result in a
compiler that optimizes final program performance.