(3 undefined) ascent to transcendence (step 1)
After many hours spent drawing trees and diagrams on paper, I have
returned to my computer elightened. I know how all the pieces go
together.

-------------------------------------------------------------------------
Thing One

`realize` has the general type:

```
SAnno GMeta Many [CMeta] -> SAnno GMeta One CMeta
```

But now is a real optimization function. It finds the least penalized
AST. The "dimensions" are 1) the `[CMeta]` list of possible
implementations for a term and 2) the `Many` that list possible
different subtrees. The `Many` appears when a term is both defined as
one or more Morloc compositions and zero or more sourced
implementations.

But while hammering this out on paper, I found there is a 3rd dimension.
The `segment` step, which I initially considered trivial, actually has
an interesting twist. In the AppS expression `AppS f xs`, where `f` is a
function and `xs` are the arguments, the arguments are all evaluated in
the parent that calls `f` and then passed to `f`. But in many cases
there is room for optimization. Argument `i` can be substituted in `f`
for the `x_i`. Essentially this is "inlining" the argument. For example:

```
foo x = f (m1 42 (g 5))
m1 y z = h 12 y z
```

Could be rewritten as by inlining the argument `z`

```
foo x = f (m1 42)
m1 y = h 12 y (g 5)
```

`m1` is a generated function, we can rearrange its arguments without
incurring unwanted side-effects.

The advantage of these rewrites is that they may circumvent unnecessary
foreign calls and serialization overhead. Suppose `f` and `m1` are in
different languages. If `g` is in the same language as `f`, then it will
usually be more efficient to evaluate it first and pass the argument to
`m1`. If, however, it is in the same language as `m1`, then it is
probably better to rewrite, since otherwise we would incurr an extra
round of serializations and a foreign call.

The `segment` function takes the full AST an breaks it into
language-specific subtrees. The subtrees are then grouped into pools,
where each subtree begins with a LamS that specifies the input arguments
(0 or more). It was initially tempting to delegate the rewrite
optimizations to `segment`, but this will not work. The function
selection optimizations done in `realize` are dependent on segmentation
(needed to determine scoring), so eventually the `realize` and `segment`
functions will need to be interwoven.

For now, however, I will ignore rewrite optimizations.

-------------------------------------------------------------------------
Thing Two

My old code was way too complicated. I can eliminate the IMeta and
Argument data structures. Explicitly threading argument data through the
tree is both complicated and unnecessary. If every segment begins with a
LamS, I can write a special map that propagates arguments down the tree.
No need to pull it out and store it.

Also I had been storing serialization information in the IMeta
structure. This also is not necessary, since I can look the data up when
I need it. This will leave more work to the `codify` function, but I
think it is better to localize the serialization task rather than pass
it in from outside.