(3 undefined) - fell into the fourth dimension
In addition to the 3 optimization dimensions discussed in 35eae0f, there
is also a (weak) optimization step needed in choosing when (and if)
serialization occurs in a pool for a given input.

All inputs to a pool start as raw JSON (i.e., packed/serialized). I can
read this raw JSON input and pass it as strings down the call tree. Then
I can deserialized when the argument needs to be passed to a concrete
function. This is how I have been handling deserialization. But there
are other options.

 1. deserialize everything in pool main
 2. deserialize at lambdas
 3. deserialize at cis applications

Arguments that are never called in the given segment, but are passed to
foreign functions, do not need to be deserialized. Actually, the fact
that they are chained uselessly through the program suggests room for
optimization, though I am not yet sure how to do this.

The main optimization goal is to minimize the number of
serialization/deserialization steps. This is trivial, to a degree, since
we can just traverse the segment to find the arguments that need to be
deserialized, then generate code that deserializes those arguments in
main and passes them as native data structures down the call tree. This
ensures deserialization occurs only once. However, if foreign calls are
made, these deserialized functions may need to be reserialized. This can
be avoided by passing both the serialized and deserialized terms. This
is ugly though.

I am not quite sure yet how deep this when-to-serialize problem goes.
There may be a trivial implementation solution that is also optimal from
a performance angle.

The how-to-pass problem, where a good solution might be able optimize
away the passing of arguments that are not used in the segment, seems
much deeper. However, here the payoffs of optimization might be small,
since we could avoid the costs of passing large objects by not passing
them at all, and rather passing keys for accessing them from local
databases/files.

For now, I will go with the naive #3, since it is easiest to implement.