(1 undefined) - wrote `serialize`, correct?
All terms are now defined in `serialize`, but I have doubts that it is
correct. See the TODO statement. I need to ensure that the arguments in
`(CMeta,[Argument])` are referring consistently to the `[Argument]` used
by the child (not inherited). I need to ensure packing/unpacking is
dealt with correctly.

Is it possible/desirable to make this a pure function (not part of
MorlocMonad)?