(5/150 fail) - rescind top-level-lambda rule
The idea that all top level expressions (i.e., ones that are first requested
by the nexus of pool) by lambdas is awkward. To make this rule hold, I
would have to add extra code to insert lambda expressions, with no
arguments, over top-level variables. On the other hand, I am doing this
anyways later, when I consult the type to determine which arguments to
pass it.