(4/150 fail) "fix" variable names
We need a permanent solution to my problem. Currently the variables
generated for source code are completely unsafe. There is no guarantee
that they do not shadow/conflict with something in the internal
language, or even with eachother. Also, variable naming is currently
independent of language. The small terms I use will be legal in most languages
(e.g., "x0", "a", "aa", "aba", etc), but I shouldn't assume that will
always be the case. Variables such as "be" and "by" would currently be
generated if enough variables are required for a function (though no
sane function would require that many positional variables).

This is not a huge concern at the moment, but in the quest to stabilize
morloc, it is pretty low hanging fruit.