(2/150 fail) identify pass-through problem
Take the code:

```
f C :: double -> double
g R :: numeric -> numeric
foo x = f (g x)
```

What is the concrete type of `foo`? The return type is the return type
of `f`: `numeric@R`. The input type is the type of `x` in the R
language. But we do not know this type. Nor do we need to. We could
express the type with the chimeric signature: `double@C->numeric@R`.

Arguments like this should be passed through a tree in serialized form
without their deserialized type ever being known.

In this commit, I've made some progress on the implementation of this,
but it is not quite working yet.