(0/152 fail) - Currying is now working
Yay. The currying itself seems to be working seemlessly and was not too
hard to implement. However, I am not happy with the state of
serialization in Morloc. Currently no polymorphism of serializers is
supported. I can't compose serializers. I want `List a` for free if a
serializer for `a` exists.

I will leave the Serializer Composition problem for the next iteration.
For now, I need to clean up my repo and push to master.