(1/157 fails) Found bug in segment
I added the fibplot example from the README and found that it breaks
strangely. The reason is nothing glamorous, it seems that I am not
passing the correct expected parent type to `segment'` at
Generate.hs:707. But fixing this problem will require a bit of tedious
rewiring.

I need to comb very carefully through the entire generator algorithm.
Lay out every cleanly. It will be a big job, probably requiring a month.
The payoff will be a cleaner mental model and a large piece of the
Morloc paper.

But this can be done after I push to master. For now I need to fix the
immediate problem. So go to Generate.hs:707 and also see the debug
output of 10_io-interop.