(14/194 fail) Fix partial ordering
Types are partially ordered over the subtype relation. I use this order
to decide which serialization function to use for a given concrete type.
A given language may have a generic serialization function with type
(`forall a . a`) as well as more specific serialization functions, such
as `forall a b -> map a b`. I always want to use the most specific
serializaer. Under the subtype relation, `forall a . a` is the smallest
type, more specific types are larger, incomparable types are undefined.

In this commit, I've fixed the subtyping for universal qualifiers and
added many tests of the desired behavior.

Before I do any more coding, the following needs to be thoroughly
documented (in the `morloc-project/docs` repo):

 - [ ] The subtype partial order definition
 - [ ] The complexity partial order definition
 - [ ] The relation between the two
 - [ ] How the subtype partial order is used to select serialization
       functions (with examples)

I should also resurrect property testing over the subtypes,
substitution, generalization and other key functions. I just fixed a bug
in the substitution algorithm, this is highly disturbing, since the
algorithm is so central to the language. I need more testing.

A prerequisite for property testing is the development of a generative
grammar over types (and eventually expressions). This grammar should
also go into the manual.