(12/194 fail) Removed space in C++ tuple type
One of my tests was failing because the compiler could not find a packer
for the type "std::tuple<$1, $2> double double". This seemed strange,
since I most definitely had this packer in my library. I found the
problem was a difference in whitespace. The C++ defaultTuple function
included space after the comma, while cppbase module did not.

This whitespace dependence is deeply disconcerting. Types should not
change with formatting style. This problem could potentially be solved
with language-specific pretty printers.

It is not the only concerning issue in the current macro-based, concrete
type system. The handling of namespaces is also awkward. Is the entire
namespace required? Is "std::tuple<$1,$2>" the same as "tuple<$1,$2>"?
This problem becomes worse when namespaces can be given aliases. Must
the Haskell "Map" type be written as "Data.Map.Map"?