Write full config to travis
The travis issues are actually revealing something meaningful this time.

When no config is given, `morloc` will use defaults for each term in the
config. But when only one line of the config was given (setting the
python interpreter), everything broke. This is because the FromJSON
instance for Config provides no defaults for the critical "home" and
"library" fields.

For now, since I am on the master branch, I won't fix these problems.
But I need to look into them when I start the next dev cycle.