(25/197 fail) Move to schema-based serialization
In schema-based serialization, I pass the data to the serialization
function along with a second input that adds type information (or
whatever). Now all serialization functions call the "pack" generic
function in the given language (and "unpack" for deserialization). What
exactly goes into the second input will vary with language.