(18/196 fail) Add Python schema-based serialization
The schema-based serialization implementation in Python is a lot cleaner
than the R one. I write the schema in the general form
(<type_name>,<type_parameters>). Examples:

 * float           --> ("float", None)
 * [float]         --> ("list", "float")
 * (float, [int])  --> ("tuple", [("float", None), ("list", "int")])
 * {a=int, b=bool) --> ("record", {a=("int", None), b=("bool", None)})

Every pack function will see an input data structure and a type with two
components (type name and type parameters). The type name is used to
find the function that will pack the data and the type parameters are
passed to that function.

I should at some point refactor the R serialization to match.