(7/199 fail) C++ translators are online
Updating the prior TODO:

 - [x] add assignment extration
 - [x] fix C++ generator
 - [ ] fix interop
 - [ ] fix nesting bug
 - [ ] test IO
 - [ ] test complex type interop

I am not happy with the elegance of the refactor. Much of the coding is
still opaque and kludgy. I should take each of my test cases, and walk
through the entire morloc program with them, describing what happens at
every stage, from parser to typechecker to generator.