(13/208 fail) Added test for new inference bug
In the script:

```
id :: Num -> Num
id C :: int -> int
y = 40;
foo = id y;
```

The `y` in `id y` should be assigned the types `Num` and `CLang int`.
Instead it is assigned only the general `Num` type. This leads to
problems in the generator where no concrete type can be found.

This appears to be a trivial wiring problem, since the program does
typecheck overall and the type off `foo` and `id y` are correct.