(4 undefined) Simplify ExprM and rewire pipeline
This commit is the outcome of a couple days of groping about for a clean
way to move from an unambiguous tree (SAnno One) to an abstract manifold
tree that represents serialization steps.

I am beginning to think that serialization state should be explicit from
the initial parsing step onwards. Serialization rules should be written
into the the typesystem.

Currently, the ExprM and TypeM grammars make serialization status
explicit, while the CType grammar does not. My current difficulties stem
from from trying to write serialization in at a late state.

Also, there is a flaw in the optimization pipeline. The language of each
function is determined in the `realize` step. From the ambiguous tree, a
single tree is selected that roughly minimizes the number of
inter-language calls. However, after conversion to ExprM, a second
optimization step uses let-syntax to reorder when and where arguments to
applications are executed. For now, optimizatio is not a primary goal,
but eventually I will need to sort this out.