(4/204 fail) Allow macros in general functions
A morloc expression such as

```
foo x = [x]
```

Calls no source function. There is no concrete language involved and no
general way to choose which language to "implement" this function in.

No language-specific pools are generated and the function `foo` is run
in the nexus. In this commit, I added handling for arguments passed to
these general functions. The outcome of calling the above function with
ANY value of x is `[x]`, where `x` is replaced with the exact input the
user provided for x. Currently no error checking is performed to ensure
that the result is proper JSON.