Convert to OrderedDict for python records
When pushing to Travis, I found the Python3 record tests were failing on
their systems while they worked on mine. This appears to be because they
are using Python 3.5 whereas I am testing on 3.8.

The problem is due to a false assumption on my part that dictionaries
are ordered. This is true only from Python 3.7 onwards. Before that,
people needed ordered dictionaries used the OrderedDict type from the
`collections` module. For maximum backwards compatibility, I will follow
suite.

Now I use OrderedDict rather than dict for schemas. I also updated to
serialization functions in pybase to lift the ordered assumption from
dictionary data. Now the keys in the OrderedDict schema are used to
order the dict data, allowing python function to use unordered dicts,
and let the morloc compiler handle ordering when needed.

The python3 unpacker returns dict, not OrderedDict, which is, I
thinkg, in accordance with what the python programmer would expect.