(6/204 fail) Fire up the failure engine
The main purpose of this dev cycle is to fix 4 github error isses:
numbers 7, 9, 10 and 11. Current tests cover 7 and 9, so I will need to
write new tests for 10 and 11.

Issue #8 is to add C++ records, which I am happy to leave for a later
cycle.

Likely, this dev cycle will not go as cleanly as planned, and new issues
may rise that I will need to address. I fear major revisions to the type
inference code may be required. I may be able to push such revisions
further down the road, but the ugliness of the current code demands a
refactor eventually.