(6/206 fail) Sidestep #11
The problem with issue #11 actually dealt with generic serialization, a
problem I do not want to address in this dev cycle. It is more likely
something that I will work on when I start working on typeclasses and
such. For now I have a hard requirement that all universal
quantification must be resolved statically.