(3/214 fail) Add new typing rule that fixes #7
Say we have a variable `x` that is defined as expression `e`. At the
point where it is declared, only a general type can be inferred for `e`.
That is, `e` is an expression that defines a data structure, but makes
no concrete function calls (e.g., `4` or `[1,2,3]`).

The concrete form of `x` cannot be deduced until it is actually used.
Then it will be cast into the language it is called from.