(2 fail) Test export with many implementations
Add test for a module that exports functions that each have
implementations in multiple languages. This is likely to be a VERY
common practice. In this case, we test an Arithmetic module that exports
add, sub, div, and mul in R, Python, and C++. The compiler will have to
choose which implementation to use in each case.

I also need to test the case where I import multiple implementations of
a function from one language. For example, several different sort
algorithms implemented in C++. Also, I need to think carefully about how
these different implementations are distinguished. The most obvious way
to distinguish them is to associate them with the source code path and
the source function name.

```
source Cpp from "foo.h" (
    "bubblesort" as sort
  , "mergesort" as sort
  , "quicksort" as sort
)

-- head doesn't currently exist, but I should eventually implement this
-- in some sort of base package, where each function is implemented in
-- every supported language. Actually, sort should be there to.
import Base (head)

export least

least xs = head (sort xs)
```

Perhaps an even better example would be the universal Base module:

```
module Base {
    import CppBase *
    import PythonBase *
    import RBase *
    ...

    export add
    export mul
    export map
    ...
}
```

This module would act as an interface, merging all implementations of
the base functions in all languages. This may include esoteric languages
or specialized languages that work only on certain architectures.

The goal of this dev cycle is to fix the bugs that currently prevent
these types of modules and to add adequate testing.