4/219 fail - test files with multiple modules
Importing a module that is defined inside the current file currently
fails. For example:

```
Module A {
    ...
}

Module Main {
    import A
    ...
}
```

The reason is simply that API.hs:pase function searches the given module
for imports and then assumes they are defined in different files, so it
then searches the filesystem without looking within the file.