(1/220 fail) add new test; find new errors
See the new multiple-instances-1 golden test. There are two errors in
the generated python code. 1) The local "foo.py" file that should be
imported as `import foo` is instead imported as `import ..foo`. 2) The
input to the manifold function which should be cast as a list of floats,
is instead uncast, remaining as a string.