(all pass) Fix bugs in direct source export
There was a serialization bug in the following case:

```
source lang from "foo.lang" (foo)
export foo
foo :: A -> B
foo lang :: a -> b
```

I've fixed this, but also found a new problem. The following does not
work:

```
source lang from "foo.lang" (foo)
export foo
foo :: [A] -> B
foo lang :: [a] -> b
```

The error is that morloc cannot serialize an existential type. The `[a]`
is initialized in the parser as an existential with a default type. The
default type SHOULD be used in serialization if no alternative is set.
This may be a simple oversight in the code ... we shall see.