Use defaults types of leaked existentials
I'm not fond of existentials leaking out of the typechecker. I would
like to add a new type such as that below:

```
data SolvedType
  = VarS TVar
  | FunS Type SolvedType
  | ArrS TVar [SolvedType]
  | NamS TVar [(Text, SolvedType)]
```

which removes the ForallT and ExistT fields and represents a solved
type. All types should be solved in the typechecker. This new type would
patch a lot of holes that appear later in the code. But I will leave
this refactor for the next dev cycle.