(2/225 fail) Test sources with declarations
Here is the Python test (there are similar ones for Cpp and R):

```
import pybase
import math

export rms

source py from "rms.py" (
    "rms1" as rms   -- (1)
  , "rms2" as rms   -- (2)
)

square x = mul x x;
rms xs = sqrt (mean (map square xs)); -- (3)

rms py :: [float] -> float;
rms :: [Num] -> Num;
```

`rms` (3) is the chosen implementation since it is in C++, which is
assumed to be faster than Python3. But the compilation dies since
`morloc` can't figure out the type of (3). This is because the C++
equations are all generic, it needs a hint that the numeric type is
`double`. I could provide a type signature for `rms`, but I'd rather not
since morloc should know, since `Num` as the default C++ type `double`.
So I want to add handling for using the default here.

If I do add a C++ type signature, a second bug occurs: the C++ code is
silently not built.