(10/238 fail) add systematic module form tests
Test combinations of 0, 1, or n imports, declarations, and/or sourced
terms. These tests are organized into the folders
`module-form-<s><i><d>`. Where `<s>` is 0, 1 or n and stands for the
number of sourced instances for the term. Similarly `<i>` for imports
and `<d>` for declarations.

The older `selection-*` and `multiple-instances-*` families of tests
should be able to be deleted now (though I should not do so until all
tests pass and then should ensure what they test truly is covered in the
`module-form-*` family).