(3/238 fail) Sort of fix multiple-instances-2-py
The realization selection is dependent on the order in which signatures
are listed in the script: this is bad. I would like to guarantee order
invariance. Maybe that can be a goal for the next dev cycle.

To recover the interesting failing state in this test, place the
signature block above the definitions of square rms.