All tests pass
In the last commit, I commented on the order dependence of instance
selection. But the real problem is that there is currently no serious
algorithm for build optimization. Implementing this build optimizer is,
of course, one of the major remaining required pieces of morloc.
However, I am happy to leave such a major project to a later dev cycle.