Start dev cycle 0.27.0
The goal of this dev cycle is to get constructors working. Something
like this:

```
Map     :: [(a,b)] -> Map a b
Map py  :: [(a,b)] -> "dict" a b
Map cpp :: [(a,b)] -> "std::map<$1,$2>" a b
```

`morloc` needs to be able to interpret the capitalized function names as
constructors. I guess the constructors can be treated specially, stored
until needed in monad stack. I should still pass them through the
typechecker, but should move them to the monad before calling the
generator, because there is no sense ever exporting them to the users
executable.