Distinguish between serialization and packing
In the past I've used "pack" and "unpack" interchangeably with
"serialize" and "deserialize". However, there is a clear difference
between them. "(de)serialize" is a function from/to JSON data, whereas
"(un)pack" is a function to/from "canonical" data that can be
unambiguously serialized.

Types such as `int`, `[(string, int)]`, or `Person {age :: int, name :
string}` can all be automatically serialized and deserialized. The user
should not have to do anything special, since the exact data type is
known and the morloc compiler understands who to work with the data
(e.g., access fields in a record or serialize a list).

However, with types like `Map a b`, where the exact data structure is
not known, the programmer needs specify how the object is made from the
data. One way to do this would be to require the programmer write a
`String -> Map a b` function. But how would they make it generic? And
also this would expose the serialization directly to the programmer and
lock morloc into particular forms. This is unacceptable. Instead the
programmer should write a `[(a, b)] -> Map a b` function. That is, given
an native input that maps uniquely to JSON, build the Map object. The
deseralization from JSON to `[(a, b)]` can be automated, and the
programmer only has to map from the pairlist to the Map object (and
there should be a single existing function for this purpose in most
languages). `[(a,b)] -> Map a b` is a "pack" operations, the reverse,
`Map a b -> [(a,b)]` is an "unpack" operation. The "pack" and "unpack"
operations are always unary functions, since the input is the
deserialized JSON data.

So, in this commit I am changing naming across the board such that
serialization is specifically `a -> Str`.

Next, I will start working on adding support for inserting pack/unpack
functions between the native types and the serial types, when needed.

I will reuse the old syntax I used for (de)serialization functions
(i.e., use the pack/unpack property in the signatures).

Here is the general deserialization scheme:

```
           .- yes --> deserialize
          /
JSON --> <canonical?>  .- yes --> deserializeRec --> constructor
          \           /
           '- no --> <data?>       .- yes --> deserialize --> pack
                      \           /
                       '- no --> <has packer?>
                                  \
                                   '- no --> DIE
```

The `data` step can be ignored for now since it will be the topic of the
next dev cycle.