(35/281 fail) (Un)pack in R with JSON type schema
The handling for this is not yet implemented in rbase, so now all R
builds are failing.

Previously I made a wonky nested list in R to describe types. But string
processing in R is atrocious, so I'd rather do it in C++. So the plan is
to generate a JSON description of the type at morloc compile time and
pass it to C++ (using Rcpp) for (de)serialization. I'm planning on using
the C++ json library nlohmann/json. This general approach can work for
any interpreted language with a good C++ binding. I'll also possibly be
able to reuse much of the C++ code between languages.

This serialization is limited to things mapping one-to-one to JSON. That
is one-to-one in the general side, not including objects in specific
languages that can be serialized to JSON; I can't use those
serializations since they are not general. Instead for them I would
first serialize to general forms (e.g., [(str, int)]) and then pass that
to a constructor to make the final native object (e.g., Map str int).