Begin dev cycle 0.28.0
The goal of this cycle is to implement import/export of type aliases.
This is useful in and of itself, but is particularly useful for
constructors. The module from which a constructor object is imported
determines which constructor functions are used. I could actually stop
at this point, push to master, and then work on refactoring the *base
modules to export the Map and Set types.