Extract Module -> SAnno building from Generate.hs
Now the conversion from `[Module] -> SAnno` happens in a dedicated
`TypeChecker/Treeify.hs` file. This simplifies the bloated Generate.hs
file and isolates all use of Module to th Parser and TypeChecker
folders.

I have some tricky refactoring coming up. The problem is that I need to
traverse the dependency graph to determine which type aliases are
defined in each module (accounting for import/export). I currently have
a hacky solution to this in Treeify. But I would rather work with a
formal DAG, where collecting all typedefs is as simple as getting the
union of the typedefs defined in the module and those inherited (and
downwards recursively).

The DAG should also simplify the wrappings in Infer.sh (I think). I will
not have to rewrite the typechecking algorithms themselves.