(5 undefined) Parser/desugar code now compiles
I am not confident the logic is correct. Next I need to define the
missing functions in the TypeChecker namespace.

I am not entirely happy with my DAG implementation. It works (probably)
but is grossly inefficient and much of the code inside DAG.hs is ugly.
There may be better existing modules out there that i could reuse.