(24/291 fail) Begin dev cycle v0.29.0
The first goal for this cycle, and possible the only goal, is to get
serialization constructors working. This will allow implementation of
Map types and other parameterized types where the parameters do not
describe the inputs.

The idea is that the user defines functions that transform the objects
into simpler forms that can be automatically (de)serialized. For
example:

```
packMap Cpp :: pack => Map a b -> ([a], [b])
unpackMap Cpp :: unpack => ([a], [b]) -> Map a b
```

These functions do not go all the way to serialization, but they reduce
one level of structure, next the a and b types would need to be handled,
and so on downwards. We will automatically build a function in whatever
language to handle the (de)serialization.