(24/293 fail) pack map by concrete type
The pack map is now of type:

```
Map (TVar, Int) [UnresolvedPacker]
```

Where TVar is the concrete type that is unpacked. The UnresolvedPacker
record contains a field storing the general type linked to the
particular unpacker. Perhaps a more suitable type would be:

```
Map (TVar, EVar, Int) [UnresolvedPacker]
```

This triple key includes the concrete type and its related general form
(along with the number of parameters in each type, which must be equal).
There is still a list of UnresolvedPacker's since I want to support the
possibility of multiple implementations or serialization paths.

However, the EVar term means that I need to know the general type in the
(de)serialization algorithm.