(10/295 fail) All C++ (de)serialization tests pass
Next need to do the same for R and Python

Some of the code is kludgy as hell. The C++ serialize and deserialize
functions are almost completely duplicated. Almost. Same with the entire
R and Python translators. I need to do some heavy refactoring to
streamline the addition of new languages.