All pass - an uneasy success
There is something still not quite right. I didn't quite implement what
I set out to implement. Instead of support for objects, I've allowed
records. I don't pass the the arguments of the record to a constructor,
though, instead I bind them into a generic structure. This means none of
the special tooling that goes into the OOP design of the other language
can be leveraged.

Any way, what this accomplishes is major and valuable. But the
serialization story is still not quite done. I still need real OOP
handling and tables.