All pass
I think the solution to the OOP problem is twofold: first records need
to be distinguished from objects. For some languages, like Haskell,
there is no distinction. But for OOP languages, object and record need
to be two different types. For one, field access may differ. In Python,
a record is accessed by field name (`d["person"]` in a `dict`), while
accessing fields in an object uses a dot (`d.person`).

On the Morloc script side, no disctinction will be made between records
and objects. Data and functions are separate. A function cannot be
serialized and cannot be put in any data structure (at least not in any
that will pass between languages).

TODO:
 [ ] field access (`person@name`)
 [ ] record arithmetic
       `Person + {money :: Int, cars :: Int} - {age}`
 [ ] object data constructors:
       `birthday p = Person { name = p@name, info = add p@age 1 }`
 [ ] allow custom record constructors