Thread NamType from parser to translator
The NamType object is passed along as the first argument to the
NamT/NamU/SerialObject type constructors. It classifies a named list as
a record, object, or table. Potentially, in the future, it could be
extended to carry additional information about records/objects/tables to
the translator.

-------------+---------------+-----------------+----------------------
             | Record        | Table           | Object
-------------+---------------+-----------------+----------------------
has field    |  yes          | yes             | no - input data may
accessors    |               | (columns)       | not be stored, can
             |               |                 | be accessed only with
             |               |                 | getter/setters
-------------+---------------+-----------------+----------------------
exact        | yes - 1-to-1  | no - depends on | no - depends on
structure    | correlation   | implementation  | implementation
represented  | with JSON     |                 |
-------------+---------------+-----------------+----------------------
can server   | yes           | no              | no
as keyword   |               |                 |
argument set |               |                 |
-------------+---------------+-----------------+----------------------

A table is parameterized with a record describing a single row, for
example:

```
table Person = {name :: Str, age :: Int}
```

This will store data isomorphic to the following record:

```
record Person = {name :: [Str], age :: [Int]}
```

In fact, the simplest way to (de)serialize a table will be to first
unpack it into a record object and then feed it into the existing
deserialization functions. Of course, there may be more efficient
solutions, especially if the table happens to represent a database
connection instead of a native structure.

Eventually I may add additional metadata to table types, such as the
uniqueness constraint used in SQL. I also need to add clean methods for
interacting with databases. Well, eventually no data should be passed as
JSON text and everything should move as binary in high-performance
databases, but that sort of optimization isn't a priority yet.