(2/298 fail) R table handling works
I'm desugaring the table type into a record immediately in the parser.
So the given type:

```
table Person = Person {name :: Str, age :: Int}
```

Will be translated to `{name :: [Str], age :: [Int]}` immediately. The
`table` annotation will be preserved in the NamType record. But since
the type is otherwise identical to a record, any machinery for dealing
with records (such as field access, arithmetic, etc) will immediately
apply to tables. Tables come with a constraint on field lengths. I
currently do not automatically generate this constraint, but I should.
I'll leave that for later, though, when I actually implement constraint
handling.

Next: extend table handling to Python and C++. I'm not quite sure how to
handle tables in each languages.

For python, a table could be a dict of vectors or a pandas DataFrame.
Certainly the latter is preferred in real data analysis.

For C++, I could use a map from column names to vectors. But for
scientific purposes, there may be better solutions, such as discussed
here: https://medium.com/@johan.mabille/xframe-towards-a-c-dataframe-26e1ccde211b

I should not make this decision at all. Instead this should be something
that is decided by the user. They would simply provide a constructor
name, e.g., `table py Person = "pandas.DataFrame" {...}`. For tables it
would be useful to have an alias over constructors. But not yet, I'll
see how things hash out in practice, then add in extra niceties. For
now, I just need to show that full generalization is possible, but can
leave actual implementation as an excersize for the reader.